Fixed SpiDma priorities, minor fixes in style.
Fixed DMA channel priorities in SpiDma class. Order of channels changed
due to the refactoring in the base GPDMA interrupt handler: interrupts
are handled in reverse order.

+++ platform/nxp/lpc11exx/clocking.c
@@ -210,7 +210,7 @@ static inline void flashLatencyReset(void)
   sysFlashLatencyUpdate(3);
 }
 /*----------------------------------------------------------------------------*/
-static void flashLatencyUpdate(unsigned long frequency)
+static void flashLatencyUpdate(uint32_t frequency)
 {
   const unsigned int clocks = 1 + frequency / 20000000;
